+++ src/lib.rs
@@ -1,2 +1,5 @@
 #[cfg(feature = "events")]
 pub use wolf_engine_events as events;
+
+#[cfg(feature = "input")]
+pub use wolf_engine_input as input;